12/28

+++ vector.h
@@ -9,13 +9,10 @@ template <typename T>
 class vector {
 private:
     T* data;         // 指向动态分配数组的指针
-    size_t capacity; // 数组的容量
-    size_t length;   // 数组中当前存储的元素数量
+    int capacity; // 数组的容量
+    int length;   // 数组中当前存储的元素数量
 public:
-    // 构造函数，初始化成员变量
     vector() : data(nullptr), capacity(0), length(0) {}
-
-    // 析构函数，释放动态分配的内存
     ~vector() {
         delete[] data;
     }
@@ -36,22 +33,22 @@ public:
     }
 
     // 重载[]运算符，提供对元素的随机访问
-    T& operator[](size_t index) {
+    T& operator[](int index) {
         return data[index]; // 返回指定索引处的元素引用
     }
 
     // 重载[]运算符的常量版本，用于const对象
-    const T& operator[](size_t index) const {
+    const T& operator[](int index) const {
         return data[index]; // 返回指定索引处的元素常量引用
     }
 
     // size函数返回数组中当前存储的元素数量
-    size_t size() const {
+    int size() const {
         return length;
     }
 
     // resize函数用于调整数组容量
-    void resize(size_t new_capacity) {
+    void resize(int new_capacity) {
         T* new_data = new T[new_capacity]; // 分配新的数组
         copy(data, data + length, new_data); // 将旧数组中的数据复制到新数组
         delete[] data; // 删除旧数组
